removed css file, was causing thumbnails on index to be bigger, used 'fluid' instead
