--- src/components/Memorial/Memorial.css
@@ -1,9 +1,0 @@
-img {
-    height: auto;
-    width: 100%;
-    
-}
-
-h1 {
-    text-align: center;
-}
